curl_ntlm_msgs.c: Fixed compilation warning from commit 783b5c3b11
curl_ntlm_msgs.c:169: warning: conversion to 'short unsigned int' from
                      'int' may alter its value

+++ lib/curl_ntlm_msgs.c
@@ -166,7 +166,8 @@ static unsigned int readint_le(unsigned char *buf)
  */
 static unsigned short readshort_le(unsigned char *buf)
 {
-  return ((unsigned short)buf[0]) | ((unsigned short)buf[1] << 8);
+  return ((unsigned short)((unsigned short)buf[0]) |
+          (unsigned short)((unsigned short)buf[1] << 8));
 }
 
 /*
